Annotations combinator: documented and unit-tested [BBPP82-307]
Annotations combination aims at reinstating brain regions
which are missing in a recent annotation file (say, one of the Mouse CCF v3 annotation files)
by falling-back to a less recent annotation file (say, one of the Mouse CCF v2 annotation files)
which has a complete set of brain regions.

* The combinate-annotations command line takes five paths as input arguments:
  - hierarchy (e.g., hierarch.json or 1.json).
  - brain-annotation-ccfv2 (e.g., path to annotation_10_2011.nrrd).
  - fiber-annotation-ccfv2 (e.g., path to fiberAnnotaton_10_2011.nrrd), required because fibers
  are missing in 2011 annotation files.
  - brain-annotation-ccfv3 (e.g., annotation_10_2017.nrrd).
  - output-path, i.e. the path where the combined annotation file will be saved.

* This change includes the scaffolding of the atlas-building-tools project (setup.py, README.rst, click CLI, tox).

Note: The structure of the project may not change drastically. Indeed, it seems that re-factoring leads
to several simplifications and files removals (towards a smaller and more focused project). Splitting and grouping
will be considered later on if the added value of separation is clear.

Change-Id: Ia3cd0c98376df43ac0f84d4039a26ff3b907cdeb