Cerebellum direction vectors: documented and unit tested [BBPP82-310]
These changes comprise

* A CLI to generate the cerebellum direction vectors:
`cerebellum-direction-vectors --annotation-path=<THE_INPUT_PATH> --output-path=<THE_OUTPUT_PATH>`.
* Unit tests

The CLI relies on:

* A util function (direction_vectors/utils.py) to compute the gradient of a scalar field after
applying a Gaussian blur. This function is instrumental in the computation of the cerebellum direction
vectors.
* A generic algorithm suited for laminar brain regions is implemented based on a blur gradient.
This algorithm is subsequently used in direction_vectors/cerebellum.py to implement the above CLI.

Note: The algorithm is the same as the original in blue_brain_atlas.
The layout of the input parameters has been changed using classes so as
to document the interface and to allow generalization.

Change-Id: Iaee8447302fe4de85363d7b438070999458776d5