Direction vectors with Regiodesics: documented and unit tested [WP10.E2.T7.S3-BBPP82-310]
Summary of changes:

* Implements a generic algorithm generating direction vectors a in 3D volume
surrounded by two regions identified as the 'bottom' and the 'top'.

* The computation of direction vectors relies on two executables of Regiodesics:
layer_segmenter and geodesics (https://bbpcode.epfl.ch/browse/code/viz/Regiodesics/tree/).
* The computation of the shared boundaries with the 'bottom' and 'top' volumes,
referred to as upper and lower shells in Regiodesics terminology, is automated.
A scipy's convolution with a neighbour filter is used to compute boundaries.

This implementation is faithful to Hugo's implementation in blue_brain_atlas.
The script which computes the direction vectors of the
mouse isocortex using the above algorithm, and possibly another one,
will be part of a separate commit.

Change-Id: I92d9aac15f316f0315cb53f307c0f444e042a0e7