Addresses issue [BBPP82-326]: overrides ccfv3 ancestors by ccfv2 [BBPP82-36]
This fixes a concern raised by Hugo Dictus while
reviewing https://bbpcode.epfl.ch/code/#/c/48332/.

The goal of this commit is to enhance the shapes of the regions
in the whole mouse brain atlas when released as a
public interactive web content.

It implements the following 'refinement' rules:

* ccfv3 voxels with zero labels are left unchanged.
* If a voxel in ccfv3 atlas is an ancestor of the corresponding voxel
in ccfv2, and if the corresponding ccfv2 label is not the label
of any voxel in ccfv3 then use the ccfv2 label instead.

The unit tests have been expanded accordingly.

Change-Id: I9d0187087c9cc5f42d8da4fdbb9e40e8dc748537