Document and test the splitting of isocortex's layer 2/3 [BBPP82-309]
This change documents and tests the algorithm of Hugo Dictus (formerly in blue_brain_atlas)
used to split the layer 2/3 of the mouse isocortex into layer 2 and layer 3.

It is based on voxel-to-layer-mesh distance computations and the knowledge of
a layer thickness ratio. The tools to compute such distances are already
part of atlas_building_tools/distances.

Note that new structure ids are created and that the input
AIBS hierarchy file (1.json) is modified accordingly.

In a nutshell:

 * A literature ratio is used to determined which voxels of layer 2/3 should sit in
    layer 2 or layer 3.

 * Voxels initially sitting in layer 2 will continue to use the same identifiers.
    Voxels sitting in layer 2/3 and identified as member of layer 3
    will use new identifiers.

Note: In the isocortex, the voxels corresponding to a name or an acronym ending with
'layer 2' are left unchanged. No voxels of the isocortex should correspond to a name
or an acronym ending with 'layer 3'.

Change-Id: I21ffb7ba3ebfdcd857935631cbcc36512e0c8df2