Average genetic markers datasets [BBPP82-313]
These changes implement operations performed on genetic markers
datasets from AIBS which are required subsequently for the computation
of cell densities.

These operations are described in the section 'Glia Differentiation' of
'A Cell Atlas for the Mouse Brain' (Eroe et al., 2018).

* Average glia subtypes datasets based on literature estimates (Zeisel et al., 2015)
to get a unique expression intensity volume for each of the 3 cell types: oligodendrocytes,
 astrocytes and microglia.

* Average cell type densities based on literature estimates (Keller et al., 2008)
to obtain the proportions of oligodendrocytes, astrocytes and microglia in the whole
mouse brain.

* Use the two previous average to derive an estimate of the overall glia density in
the whole mouse brain.

The computation of average densities differs from the one in blue_brain_atlas,
see discussion in the comments of https://bbpteam.epfl.ch/project/issues/browse/BBPP82-313.

Change-Id: I0dd54e585b820e71e1e613a06cb5150bb25b7978