Fixes bug with assignment of subregion direction vectors to whole region [BBPP82-306]
Previously, the copy of direction vectors computed on a subregion of
the Isocortex, e.g., AIp, was done for every multi-index in
the Axis Aligned Bounding Box of this subregion.

This is not correct, since this writes also on neighbouring subregions.
The subregion mask has to be used to restrict the scope of the
copy.

In addition, the algorithm wasn't handling missing top or bottom parts,
which can cause Regiodesics to crash (this happens with CCFv2's FRP subregion).
This commit addresses this situation by skipping incomplete iscortex regions
(i.e., regions which don't intersect the layer 1 or the layer 6) and issuing
a warning. Direction vectors which cannot be computed are assigned the value
(NaN, NaN, NaN).

Eventually, the command writing the distance report was flawed (wrong mode)
and is corrected by this change.

Note: Besides the bug fix, the file README.rst has been updated
with installation instructions for the new C++ dependencies.
Hence this is also a documentation fix.

Change-Id: I8ca3e22ecf363e16c707d0947ec4a3767f69d855