Compute densities of different cell types, including glia cells, the inhibitory and excitatory cells [BBPP82-312]
This commit is the conclusion of the Jira ticket [BBPP82-312].
It re-implements, documents and tests the computation of cell densities
the same way it was previously computed in blue_brain_altas/create_density.py.
It also places a CLI on top of it.

Lexicon:
A density value is float number corresponding to the number of cells in a voxel.
A density field is a 3D volumetric array assigning to each voxel a density value.

By means of the CLI, you can compute and save the following mouse brain cell
densities under the form of density fields.

* overall cell density
* overall glia cell density and overall neuron density
* among glial cells:
    - astrocyte density
    - oligodendrocyte density
    - microglia density
* among neuron cells:
    - inhibitory cell density
    - excitatory cell density

Note: Although some consistency checks are ensured by unit tests,
higher level validation criteria are missing.
I suggest the following strategy if further changes are to be considered:
state all validation criteria beforehand and articulate the density
creation process as a well-defined optimization problem under constraints.

Change-Id: Ie1329c6c5cb9b8002e89b4e3fced3e3ea143b081