Document and test cell detection and average soma computation [BBPP82-312]
This commit re-implements, documents and unit-tests the algorithm of
blue_brain_atlas/create_densities
which computes the average radii of somas in different mouse brain
regions based on the analysis of Nissl stains and annotated brain slices.

In few words, somas are detected as local intensity maxima of a color
inverted image and a disk-shape fitting algorithm computes a radius
estimate for each detected soma.

The generated average radii estimates are optionally integrated in the
subsequent computation of cell densities. Radii are used as a correcting
factor associated to each region where it is possible to compute them.

The final computation of cell densities (glia, overall cell and neuron
densities, inhbitory and excitatory cell densities) is the subject of the
next commit: https://bbpcode.epfl.ch/code/#/c/49592/

Note: the execution time of this algorithm is remarkably long.

Change-Id: I04eca7b27db2968d14523b21c4b309c29e74e4dc