[FIX] Fixes wrong input for VoxelData constructor when saving Cerebellum direction vectors
This bug has been reported by Jonathan Lurie and Nabil Alibou (DKE Team):

The script generating the direction vectors of the mouse Cerebellum
crashes at the moment of saving the direction vectors array to disk.

The issue is solved by removing the extraneous (and erroneous)
VoxelData object around the array that is returned by the internal
method.

Besides, the other following fix has been done: the Isocortex direction
vectors are now saved to disk with the type float32.

There were supposed to be saved that way but the type specification was
actually missing. (Thanks to Jonathan for reporting this!)

Change-Id: Ib0623c1214048f3b686a499ad6f3378cd0293b60