Document and test the generation of cell positions [BBPP82-314]
This change documents, tests and re-implements the creation of cell positions
originally written in blue_brain_atlas and described in "A Cell Atlas for the
Mouse Brain" by C. Eroe et al., 2018.

The algorithm takes a series of volumetric density files as input, i.e.,
a series of nrrd files where each voxel bears a float value which represents
the number of cells lying inside it.

Each input density file corresponds to a different cell type taken among
astrocyte, microglia, oligodnedrocyte, excitatory and inhibitory.

Each density file is used to produce both a target cell count
(the rounded density sum) and a probability distribution over the mouse brain voxels.
An acceptance-rejection method is then applied to select voxels where
cells will be placed. Inside each selected voxel (a voxel can be selected multiple times),
 a random 3D point is choosen according to a 3D uniform distribution.

The output is an h5 file containing a float32 numpy array of shape (N, 3)
where N is the total number of mouse brain cells. The 3D coordinates of
each generated cell position are expressed within the 3D orthonormal frame of the
 associated mouse brain annotated volume (hence the offset and the voxel dimensions
 are taken into account).

Note: The CLI takes as input a yaml configuration file holding the paths of
the density files to be used. This is similar to the interface choosen for the
combination of genetic markers, see https://bbpcode.epfl.ch/code/#/c/49082/.

Change-Id: Idfdfa0fe867d7aac8968236d5beb75c024b96a6a