Flatenning a layered region by collapsing direction vectors streamlines [NSETM-1035]
This change implements a flatmap following the lines of
https://bbpcode.epfl.ch/code/#/c/49332/.

Flatmaps are used by nse/white-matter-projections to create long-range
connections.

Description of the flattening algorithm:

The brain region to be flattened is assumed to be laminar, i.e., foliated by layers.
The brain region is bundled with its 3D direction vectors whose streamlines are used to
create a voxel-to-pixel mapping in the following way.
For each voxel, a streamline passing through it is drawn. Here we call a streamline a polygonal
line following the stream of the discrete `direction_vectors` field.
The streamline intersects the boundary surface shared by `layer_1` and `layer_2`.
The map assigning intersection points to voxels is finally post-composed with an authalic
transformation, i.e., a transformation flattening the boundary surface while minimizing locally
area distorsion.

Note: the code relies on the CGAL python bindings of
https://github.com/mgeplf/cgal-pybind/tree/authalic
and the companion commit https://bbpcode.epfl.ch/code/#/c/52270/
which is a python-C++ binding for the computation of streamlines
intersections.

Change-Id: I646104880e83d2a149da9b875e83da51378884bb