(DOC FIXES) Additions to README.rst and to several module top docstrings
This change aims at easing off the comprehension of this
python project.

The context of the Cell Atlas creation is recalled.
The main functions of the project are outlined at the top
of README.rst.

In addition, several typos have been fixed.
A FIXME comment warns us that microglia cell densities can take negative values.

Change-Id: Ib3d112db36e86fa04b64a42077708a8c4cddee0a