Fix isocortex direction vectors computation [BBPP82-440]
The inclusion of layer 6a as the bottom region used for
direction vectors computation by Regiondesics has a negative impact
on the number of voxels bearing a reliable distance information for
cell placements (more unreliable placement hints are generated).

Indeed, after this change, which is not part of the initial implementation
of Hugo Dictus in blue_brain_atlas/compute_orientation/isocortex.py,
the number of ray-mesh interesections with the mesh of the right hemisphere
of the isocortex has decreased: 89% of hits versus the expected 97%.

This impacts directly the statistics contained in the generated distances_report.json.

Removing layer 6a from computations solves this issue.
The reason why layer 6a should be excluded is simple:
layer 6a is not part of the bottom boundary of the isocortex with its complement
in the mouse brain. Layer 6b alone is a much better representative of this
boundary for Regiodesics.

Change-Id: I9a991146c3634484d67cbfc22647fc58f87fa1ac