(FIX) Interpolate distance information of problematic voxels with the distance information of valid voxels
In the context of the computation of placement hints, a voxel is said to
be problematic or invalid, if it failed one of several tests (mostly consistency checks)
for the computted distance information (distance wrt to direction vector from the voxel
to a layer boundary).

One of these tests for the isocortex, is the comparison of the local thickness computed
at a voxel via [PH]_layer_i.raw[1] - [PH]_layer_i.raw[0] against an upper bound which is
an anatomical reported value plus some tolerance (twice the edge length of a voxel).

Until now, the distance information of problematic voxels was re-computed as an interpolated value
of valid distances of neighbours only when the problematic distance is NaN.

With this change, the interpolation happens for any problematic voxels.
We expect that this change will result in more reliable local thickness information, but
this needs to be validated.

Note: From now on, two reports are created and merged into a single one.
The first report describes errors detected before interpolation
of invalid distance information.
The second one describes the issues remaining after interpolation.
The saved problematic volume is a uint8 mask for the problematic voxels detected before and
after interpolation. The voxel value 0 means that no problem was detected.
The code 1 means that a problem was detected before interpolation but not after.
The code 2 means that a (excessive thickness) problem persists after interpolation.

Change-Id: I253d7d85322c3d9667e14b791cf86f2455386f35