(FIX) Create new identifiers for both layer 2 and layer 3 [NSETM-1288]
This change aims at fixing two issues in the splitting process of the layer
2/3 (AIBS 1.json) of the mouse isocortex:

(1) Currently, regions whose names end with "2/3" can re-use
their former identifiers to create a sub-region in layer 2.
We shall prohibit this and we shall keep each node whose name ends with
"2/3" in the hierarchy. The children list of each 2/3 leaf node will be
populated with two new nodes, one for layer 2 and one for layer 3.

(2) If a region name is not represented inside the annotated volume
by any voxel, the modification of the hierarchy file is quite unexpected:
the ending "2/3" is replaced by "2" but no new node is created for
the layer 3. We shall prohibit this and create two new nodes on every turn,
one for layer 2 and one for layer 3.

This change simplifies and clarifies the splitting process in decoupling
the modification of the hierarchy file from the modification of
the annotated volume.

The hierarchy file is first modified in a straightforward way,
independently of the annotated volume: add to every node whose
"name" ends with "2/3" (this always corresponds to a leaf node in AIBS 1.json)
two children with names ending with "2" and "3" instead.
The created children are assigned distinct and new identifiers.

The annoted volume is then modified, using the new identifiers of the
modified hierarchy.

The algorithm splitting the volume is left unchanged: we compute
approximate distances from voxels to the top and to the bottom of the
layer 2/3 volume. Then we use a constant thickness ratio to decide whether
a voxel belongs to layer 2 or layer 3.

Note: When creating a new child node in the hierarchy file, we set
its "parent_structure_id" in a consistent way. All the other attributes are
copies of the parent attributes. It is unclear what the value of "st_level"
and "atlas_id" should be,
see https://community.brain-map.org/t/what-is-the-meaning-of-atlas-id-and-st-level-in-1-json/895

Change-Id: I7a3fe5df2e4634f81c07e609e96a56143fe7a369