Read compilations of density-related measurements and save them under a uniform format [NSETM-1376]
This change implements a CLI which reads two measurement compilations:

* mmc3.xlsx from Kim et al., 2017 (PV+, SST+ and VIP+ densities)
* gaba_papers.xlsx, a measurement compilation of Dimitri Rodarie (BBP)

All measurements are saved under a single CSV file thanks to the CLI.
The format of the file is described when specifying the --help
option of the command line:
atlas-building-tools cell-densities compile-measurements --help

A companion CSV file containing the list of regions with homogenous neuron type
is also saved for later use in the densities module.

Note: the output format
* should ease off any subsequent extensions of the series of measurements
* should ease off the subsequent operations on measurements made
in the densities module (refinement of inhibitory neuron densities
envisioned in https://bbpcode.epfl.ch/code/#/c/49734/ by Dimitri Rodarie)

The CLI should be abandonned once the series of measurements is saved
permanently in Nexus is a possible better format.

Note: only NaN measurements are filtered out. Measurements with a region
name which is not a valid AIBS region name are kept as they are. They need to be
filtered out when operating on AIBS annotated volumes.

Change-Id: I0942d6dfa371ee08b60bd2c685e2b26b91a1e949