Compute the direction vectors of the AIBS mouse thalamus [NSETM-1417]
This changes implements the computations of direction vectors
for AIBS mouse thalamus as initially designed in
https://bbpcode.epfl.ch/code/#/c/48716/ by Hugo Dictus (BBP).

The computation relies on the simple gradient blur algorithm:
we define two scalar fields, one for the reticular nucleus of
the thalamus (RT), the other for its complement inside the thalamus (TH)
. Source and target regions are defined by using the opposite hemisphere
and the common outer boundary of RT and TH.

The gradient of the fields obtained after blurring the above two fields
provides us with the desired 3D vector field.

See
https://bbpteam.epfl.ch/project/issues/browse/NSETM-1417 for details and
pictures.

Change-Id: Ie136a534753a0ad324005c47d612fe63bcd4918f