Computes the placement hints of the mouse thalamus [NSETM-1433]
This change implements the computation of the placement hints for the
mouse AIBS thalamus as originally written in
https://bbpcode.epfl.ch/code/#/c/53190/ by Hugo Dictus.

The algorithm is similar to the one used for the isocortex or CA1.
The thalamus is handled as a 2-layer region, the first layer of which
is the reticular nucleus of the thalamus.

As the bottom of the thalamus is too uneven to obtain meaningful
ray-mesh intersections, each thalamus hemisphere is replaced by
its convex hull, which provides us with a smooth approximation.

Change-Id: I7746703d8a7934a4978a04c35323efd13d5328b7