Removes the dependency on nse/atlas-analysis [NSETM-1454]
The nse/atlas-analysis project adds dependencies to
atlas-building-tools such as vtk and geomdl, which are not used
by atlas-building-tools.

In order to create a spack package for atlas-building-tools,
spack packages for the latter dependencies should also be created.

With a view to take a shorter path to packaging, we consider
removing the dependency on atlas-analysis with this change.

In addition, this change turns cairosvg into an "extras" dependency:
pip will try to install cairosvg only if the extras_require option
"cell-detection" is specified. The cell_detection module is the only
module depending on cairosvg. At the moment, there is not spack package
for cairosvg on BB5 (although several cairo packages are available).

Note: this change duplicates several functions of nse/atlas-analysis
to make it independent of the latter.

Change-Id: Ie65848168ec0ec11409dca99af60e0e4485548c7