Report consistency issues with placement hints and use interpolation as a fix [NSETM-1343] [NSETM-1413]
This change implements

- three additional consistency checks for generated placement hints
as originally written in https://bbpcode.epfl.ch/code/#/c/48301/ by Hugo Dictus.

- the interpolation of additional inconsistent distances as originally written in
 https://bbpcode.epfl.ch/code/#/c/48716/  by Hugo Dictus.

The newly reported issues are:

* distances wrt the direction vector of a voxel are not
sorted consistently with the layer numbering

* distances wrt the direction vector of a voxel to the boundaries of the
voxel layer are not consistent (distance to layer top < distance to
layer bottom).

* a non-positive thickness of layer 1 (at the top, the shallowest layer)
is inferred from the distances associated to a voxel

The interpolation of invalid distances by valid ones is performed as before, but
with the following change: valid values are only taken in the hemisphere and
in the layer of the voxel whose invalid distance value is interpolated.

Change-Id: I8b4f7d0d3cb0a644659c9af72e89e32848b3f661