(FIX) Fix errors with the cell-densities CLI [NSETM-1463]
This changes fixes the following errors reported by Nabil Abilou (DKE):

* atlas-building-tools cell-densities glia-cell-densities
The values of a dict were queried with "values" and not "values()".

* atlas-building-tools cell-densities inhibitory-neuron-densities
The input neuron count had an unexpected float type.
Corrected by setting the expected int type for every region cell counts.
Doc strings indicate the type to use.

* a warning is issued by numpy-quaternion, telling us that
numba is not installed and hence cannot be used to speed up
computations. This is solved by pinning the numba version to
0.48.0.

The above two CLIs are now unit-tested.

Important: the CLI named inhibitory-neuron-densities has been renamed
inhibitory-and-excitatory-neuron-densities

Change-Id: I056c6a2e7edc59fac4ecab168d10ef8605264659