Interpolate invalid direction vectors with valid ones [NSETM-1437]
This change implements the interpolation of the direction vectors
as orignally written in https://bbpcode.epfl.ch/code/#/c/48716/.

A single command line has been added:

- atlas-building-tools direction-vectors interpolate <OPTIONS>

With the first option, i.e., --nans, the command interpolates
[NaN, NaN, NaN] direction vectors by non-[NaN, NaN, NaN] direction
vectors of neighbouring voxels.

With the second option, i.e., --mask-path, the command interpolates the direction
vectors of the voxels specified by means of a mask (nrrd file where each voxel
with non-zero value is filtered-in). It uses non-[NaN, NaN, NaN] direction
vectors of nearby voxels out of the mask as the known values of the interpolation.

This change extends the use a metadata json file to describe a brain region and
its layers. The code has been refactored make this usage uniform accross several
functions and CLIs. In particular, the CLI of atlas_building_tools/app/flatmap.py
has been changed.

TODO: A further refactoring is still needed so that a metadata file / object is
also used to instantiate a LayeredAtlas (see atlas_building_tools/placement_hints).

Change-Id: If78cd72178c1df8ab96ec5a8f7fecf65358cc4ef