Lower requirements in setup.py [NSETM-1454]
The goal of this change is to lower atlas-building-tools
requirements so that the corresponding spack
package py-atlas-building-tools can use the existing spack
packages for those requirements, without any modification.

The versions specified in setup.py have been modified according
to what is available is the BBP spack repository.
Tests indicate that the current code is compatible with these
lower versions.

See https://github.com/BlueBrain/spack/pull/1152 for full context.

After this commit, a relase is required in order to resolve the above
PR issues.

Note: It might be necessary to also lower the trimesh requirement (>=3.6.18)
in common/cgal-pybind, since otherwise atlas-building-tools will not
be able to use cgal-pybind. After a new release, cgal-pybind>=0.0.2
should be changed to cgal-pybind>=0.0.3.

Change-Id: I8bf49df9364fb7987009934293711a846c82e779