mtype densities: better support for mapping.tsv files [NSETM-1487]
This change lets users specify a mapping.tsv file containing no
mtype with inhibitory synapse class.

In addition, it documents the option --placement-hints-config-path
of the command line

atlas-building-tools cell-densities mtype-densities

An exception is raised if the keys of layerPlacementHintsPaths in
the provided yaml file are not of the expected form, i.e.,
either "y" of "layer_<index>" for a 1-based index.

Change-Id: I372ec3720c88d53962500166fe6005e615d6442f