Refactoring: broader use of metadata.json, improve placement hints report [NSETM-1474]
This commit changes:

- (1) the default input defining a region and its layers that is
used to create placement hints of the mouse isocortex, CA1 and thalamus

- (2) the report on problematic distance computations created during
the computation of placement hints.

For (1), we don't specify anymore an input brain region by means of an acronym
and a list of regular expressions (layer definitions). Instead,
we use the metadata json files stored in app/data/metadata which contains
the required information displayed in a format that we would like to be "standard".
This is enables us to handle the Thalamus like any other regions.

For (2), we have removed the so-called "obtuse angles" problem from the distance
report created after interpolation, because interpolation has no effect on direction
vectors and the associated problematic voxels cannot be checked again.
As a result, the proportion of voxels with at least on distance-related problem
appearing in the report written after interpolation is only the accumulation of
the other problems listed in this report.

Change-Id: Iafaaf1af20b2f6a28ce6ed361c3b390cafbf3727