Ensure cell inhibitory neuron densities consistency [NSETM-1506]
This change implements the final step of the computation
of volumetric inhibitory neuron densities as originally
written in https://bbpcode.epfl.ch/code/#/c/49734/ by Dimitri
Rodarie.

The average density estimates obtained from the scientific literature
or from a linear fitting based on the latter are modified so
that for every region where an estimate is available, we have:

- gad67+ average density >= sum of average densities of pv+, sst+ and piv+
- gad67+ average density <= average neuron density; the right hand side  comes from
the overall volumetric neuron density which has been estimated in an earlier step.

Whenever possible, the modification keeps the estimate in a range based on the standard
deviation of the initial estimate.

The final average density estimates are used to create density volumetric files
for the cell types "gad67+" (inhibitor neurons), pv+, sst+ and vip+.

Change-Id: Id2bcc8792a62ecd6f00e12b4121c8ae43d688005