Replace noisy warning by a note in the main CLI help message [NSETM-1513]
The warning issued when cairosvg is not installed has been replaced by
a note in the main CLI help message.

This way, the information about the disabled commandline is
only displayed when the user types the main
`atlas-building-tools --help` command.

The information is then printed before the list of
available subcommands.

Change-Id: I3deb1f712f6ea9da59918fc4007cc00fe387a42d